+++ test/hop/integration_tests.dart
@@ -1,0 +1,19 @@
+part of test_hop;
+
+class IntegrationTests {
+  static void run() {
+    test('hop output is sorted', _testOutputSorted);
+  }
+
+  static void _testOutputSorted() {
+    final onComplete = expectAsync1((Future<io.ProcessResult> f) {
+      expect(f.hasValue, isTrue);
+      final pr = f.value;
+      final lines = pr.stdout.trim().split('\n');
+      expect(lines, orderedEquals(['about', 'docs', 'test']));
+    });
+
+    final f = io.Process.run('hop', [Runner.RAW_TASK_LIST_CMD]);
+    f.onComplete(onComplete);
+  }
+}